Ajustes na arquitetura do projeto e implantaçao do componente de  unidades
