+++ src/components/equipeCard.js
