+++ styles/globals.css
@@ -7,3 +7,9 @@
 html {
   scroll-behavior: smooth;
 }
+* body{
+  margin: 0;
+  padding: 0;
+  border: 0;
+  box-sizing: border-box;
+}
